moved Monitoring to core, closes #100
